+++ ssi-vc/Cargo.toml
@@ -44,7 +44,7 @@ k256 = { version = "0.11", features = ["ecdsa"] }
 keccak-hash = { version = "0.7" }
 serde_jcs = "0.1"
 ssi-crypto = { path = "../ssi-crypto" }
-ssi-ldp = { path = "../ssi-ldp", features = ["aleo", "example-http-issuer"] }
+ssi-ldp = { path = "../ssi-ldp", features = ["aleo", "example-http-issuer", "secp384r1"] }
 ssi-dids = { path = "../ssi-dids", version = "0.1", features = ["example"] }
 josekit = "0.8.2"
 time = "0.3.20"
