+++ tests/issuer-http-5678.json
@@ -11,6 +11,20 @@
       "controller": "https://vc.example/issuers/5678",
       "type": "Multikey",
       "publicKeyMultibase": "z6MkrJVnaZkeFzdQyMZu1cgjg7k1pZZ6pvBQ7XJPt4swbTQ2"
+    },
+    {
+      "@context": "https://w3id.org/security/multikey/v1",
+      "id": "https://vc.example/issuers/5678#zDnaepBuvsQ8cpsWrVKw8fbpGpvPeNSjVPTWoq6cRqaYzBKVP",
+      "controller": "https://vc.example/issuers/5678",
+      "type": "Multikey",
+      "publicKeyMultibase": "zDnaepBuvsQ8cpsWrVKw8fbpGpvPeNSjVPTWoq6cRqaYzBKVP"
+    },
+    {
+      "@context": "https://w3id.org/security/multikey/v1",
+      "id": "https://vc.example/issuers/5678#z82LkuBieyGShVBhvtE2zoiD6Kma4tJGFtkAhxR5pfkp5QPw4LutoYWhvQCnGjdVn14kujQ",
+      "controller": "https://vc.example/issuers/5678",
+      "type": "Multikey",
+      "publicKeyMultibase": "z82LkuBieyGShVBhvtE2zoiD6Kma4tJGFtkAhxR5pfkp5QPw4LutoYWhvQCnGjdVn14kujQ"
     }
   ]
 }
